Add Light Component, Light Manager

+++ AoEngine/Source/AoMesh.cpp
@@ -2,7 +2,7 @@
 #include "AoApplication.h"
 #include "AoRenderer.h"
 
-AoMesh::AoMesh( const string& Name ) 
+AoMesh::AoMesh( const string& Name )
 	: Name( Name )
 {
 }
@@ -19,11 +19,11 @@ void AoMesh::Initialize( VertexVector&& Vertices, IndexVector&& Indices )
 
 	if ( Device != nullptr )
 	{
-		VertexCount = Vertices.size( );
+		VertexCount = static_cast< uint32 >( Vertices.size( ) );
 		VertexSize = sizeof( AoGenericVertex );
 		D3D11_BUFFER_DESC VBD;
 		VBD.Usage = D3D11_USAGE_IMMUTABLE;
-		VBD.ByteWidth = static_cast<unsigned int>( VertexCount * VertexSize );
+		VBD.ByteWidth = static_cast< unsigned int >( VertexCount * VertexSize );
 		VBD.BindFlags = D3D11_BIND_VERTEX_BUFFER;
 		VBD.CPUAccessFlags = 0;
 		VBD.MiscFlags = 0;
@@ -31,11 +31,11 @@ void AoMesh::Initialize( VertexVector&& Vertices, IndexVector&& Indices )
 		VBInitData.pSysMem = &Vertices[ 0 ];
 		Device->CreateBuffer( &VBD, &VBInitData, &VB );
 
-		IndexCount = Indices.size( );
+		IndexCount = static_cast< uint32 >( Indices.size( ) );
 
 		D3D11_BUFFER_DESC IBD;
 		IBD.Usage = D3D11_USAGE_IMMUTABLE;
-		IBD.ByteWidth = static_cast<unsigned int>(sizeof( unsigned int ) * IndexCount );
+		IBD.ByteWidth = static_cast< unsigned int >( sizeof( unsigned int ) * IndexCount );
 		IBD.BindFlags = D3D11_BIND_INDEX_BUFFER;
 		IBD.CPUAccessFlags = 0;
 		IBD.MiscFlags = 0;
